module_gmail_sender/requirements.txt
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.24.0
